Rename CameraUseCase to CameraSystem (#393)
This commit renames CameraUseCase to CameraSystem to avoid confusion with the "UseCase" pattern from Clean Architecture. The previous name was misleading, as the class did not follow that specific architectural pattern.

The new name, CameraSystem, more accurately reflects the class's role as the primary, high-level interface for interacting with the device's camera hardware and subsystems.

GitOrigin-RevId: 44624ac694dcb4d7b58d5951a62a77925483564e
Change-Id: If320ca8c1fed3ff8fbd4830c0fb0c27483815faa