Correctly apply semantics in QuickSettingsScreen (#420)
The semantics modifier in QuickSettingsScreen was being called within an
`apply` block on the Modifier. This was incorrect because the `semantics`
function returns a new Modifier instance, which was being discarded as it
was not the return value of the `apply` block.

As a result, the `stateDescription` for the quick setting was never
actually applied to the composable.

This change moves the `semantics` call to be a direct part of the
Modifier chain. This ensures that the Modifier with the correct
semantics is used, and the `stateDescription` is properly set.

This fixes the
`ConcurrentCameraTest#concurrentCameraMode_whenEnabled_disablesOtherSettings`
test, which relies on checking the `stateDescription` to verify that
other settings are correctly disabled when concurrent camera mode is
enabled.

GitOrigin-RevId: db2f691235077401abd68fb590f72579db82b7de
Change-Id: I8f1a108457e88db733daac78198d785f83123cd8